requirements.txt
fastapi
uvicorn
httpx
python-dotenv
supabase
python-jose[cryptography]
bcrypt
